[Feat/#7] VC 코드 분리 및 Base 파일 활용
NaverShoppingListView, ShoppingListCollectionViewCell
